+++ components/ADempiere/Form/VFileImport/index.vue
@@ -1,19 +1,20 @@
 <!--
-ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
-Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A.
-Contributor(s): Elsio Sanchez elsiosanchez15@outlook.com https://github.com/elsiosanchez
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
+ ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
+ Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ Contributor(s): Elsio Sanchez elsiosanchez15@outlook.com https://github.com/elsiosanchez
+ Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
 
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <https:www.gnu.org/licenses/>.
+ You should have received a copy of the GNU General Public License
+ along with this program. If not, see <https:www.gnu.org/licenses/>.
 -->
 
 <template>
@@ -58,7 +59,8 @@ along with this program. If not, see <https:www.gnu.org/licenses/>.
             </el-button>
           </template>
         </selectTable>
-        <uploadFile
+
+        <select-file
           v-if="currentSetp === 2"
         >
           <template v-slot:footer>
@@ -105,7 +107,8 @@ along with this program. If not, see <https:www.gnu.org/licenses/>.
               </el-button>
             </span>
           </template>
-        </uploadFile>
+        </select-file>
+
         <saveProcess
           v-if="currentSetp === 3"
         >
@@ -140,7 +143,7 @@ import store from '@/store'
 
 // Components and Mixins
 import selectTable from './selectTable.vue'
-import uploadFile from './uploadFile.vue'
+import SelectFile from './SelectFile'
 import saveProcess from './saveProcess.vue'
 // Utils and Helper Methods
 import { isEmptyValue } from '@/utils/ADempiere'
@@ -153,7 +156,7 @@ export default defineComponent({
 
   components: {
     selectTable,
-    uploadFile,
+    SelectFile,
     saveProcess
   },
